bigdecimal

+++ src/main/java/com/tempus_api/dtos/EmployeeDto.java
@@ -1,10 +1,11 @@
 package com.tempus_api.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record EmployeeDto (String name,
                            String password,
-                           double earningHour,
+                           BigDecimal earningHour,
                            UUID enterpriseId
                            ){
 }
